Add module Authenticatable
