+++ app/models/user.rb
@@ -1,4 +1,5 @@
 class User < ApplicationRecord
+  include Authenticatable
   attr_accessor :remember_token
 
   VALID_EMAIL_REGEX = /\A[\w+\-.]+@[a-z\d\-]+(\.[a-z\d\-]+)*\.[a-z]+\z/i
@@ -15,28 +16,4 @@ class User < ApplicationRecord
 
   validates :password, presence: true, length: { minimum: 6, maximum: 200}
 
-  # Returns the hash digest of the given string.
-  def User.digest(string)
-    cost = ActiveModel::SecurePassword.min_cost ? BCrypt::Engine::MIN_COST :
-               BCrypt::Engine.cost
-    BCrypt::Password.create(string, cost: cost)
-  end
-
-  def User.new_token
-    SecureRandom.urlsafe_base64
-  end
-
-  def remember
-    self.remember_digest = User.new_token
-    update_attribute(:remember_digest, User.digest(remember_token))
-  end
-
-  def authenticated?(remember_token)
-    return false if remember_digest.nil?
-    BCrypt::Password.new(remember_digest).is_password?(remember_token)
-  end
-
-  def forget
-    update_attribute(:remember_digest, nil)
-  end
 end
